Fix @_cdecl parameter marshalling: UTF-8 strings, collection pointer, nint length
Three NativeAOT device crash fixes in @_cdecl wrapper parameter marshalling:

1. String params in enum case/subscript wrappers: SwiftString.Buffer struct is
   not NativeAOT-marshalable. Switch to UTF-8 ptr+len (IntPtr + nint) matching
   the proven property wrapper pattern. Swift reconstructs via String(bytes:encoding:).
   Fixes Reachability FailedToCreateWithHostname and KeychainAccess subscript crashes.

2. Collection params (Array/Dictionary/Set) in @_cdecl constructors: PayloadBuffer.Buffer
   dereferences to get the VALUE, but Swift .load(as:) needs a POINTER to the value.
   Use Payload.DangerousGetHandle() for @_cdecl wrappers instead.
   Fixes CryptoSwift HMAC(byte[]) 6.5GB allocation crash.

3. UTF-8 length uses nint (not int) to match Swift Int (64-bit on ARM64),
   preventing potential truncation on very large strings.

+++ src/Swift.Bindings/src/Emitter/StringEmitter/Handler/MethodSignature.cs
@@ -517,12 +517,12 @@ namespace BindingsGeneration
                 var csParamName = NameProvider.GetCSharpParameterName(argument);
 
                 // @_cdecl property wrapper: String params use UTF-8 pointer + length.
-                // Must match the P/Invoke signature (IntPtr {name}Utf8Ptr, int {name}Utf8Len).
+                // nint matches Swift's Int (64-bit on ARM64) to avoid truncation.
                 if (_env.MethodDecl.UsesCdeclPropertyWrapper &&
                     argument.SwiftTypeSpec is NamedTypeSpec cdeclStrArgNts && cdeclStrArgNts.Name == "Swift.String")
                 {
                     AddParameter("IntPtr", csParamName + "Utf8Ptr");
-                    AddParameter("int", csParamName + "Utf8Len");
+                    AddParameter("nint", csParamName + "Utf8Len");
                     continue;
                 }
